[userpages StreamPage/SecurityView] Factor out Slider code into new files.
