+++ app/src/userpages/components/StreamPage/Show/SecurityView/securityView.pcss
@@ -72,78 +72,3 @@
 .SecurityOptions.encrypted {
   color: #2AC437;
 }
-
-.SecuritySlider {
-  pointer-events: none;
-  position: absolute;
-  width: 100%;
-  left: 0;
-  right: 0;
-}
-
-.SliderTrack {
-  position: absolute;
-  width: 100%;
-}
-
-.SliderTrack.SliderProgress {
-  transition: all 0.3s;
-  width: 0%;
-}
-
-.SliderTrack::before {
-  /*
-   * Use pseudo element for track background
-   * otherwise vertical center against TrackStops is off by a 0.5px
-   */
-  background: #D8D8D8;
-  position: absolute;
-  content: ' ';
-  top: 0;
-  left: 0;
-  transform: translateY(-50%);
-  height: 2px;
-  width: inherit;
-}
-
-.SliderTrack.SliderProgress::before {
-  background: currentColor;
-}
-
-.SliderStop {
-  position: absolute;
-  width: 10px;
-  height: 10px;
-  border: 1px solid #D8D8D8;
-  border-radius: 100%;
-  background-color: #D8D8D8;
-  transform: translate(-50%, -50%);
-  z-index: 1;
-
-  &.highlighted {
-    background-color: white;
-    border: 1px solid currentColor;
-    transition: all 0.1s;
-  }
-}
-
-.SliderThumb {
-  position: absolute;
-  right: 0;
-  top: 0;
-  transform: translate(50%, -50%);
-  background: white;
-  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.17);
-  height: 32px;
-  width: 32px;
-  border-radius: 100%;
-  color: currentColor;
-  z-index: 1;
-
-  .SliderStop,
-  .SliderStop.highlighted {
-    left: 50%;
-    top: 50%;
-    background-color: currentColor;
-  }
-}
